Enclosure Crane Concept Review
I gave this presentation to a collection of my mentors and fellow engineers at the W. M. Keck Observatory early in my internship to narrow down and dissect several design concepts for a lift fixture within our adaptive optics enclosure. The presentation was a great learning experience for understanding the collaborative side of the design process and honing my critical thinking skills as an engineer.